+++ inst/d3js/modelStudio.js
@@ -14,7 +14,8 @@ r2d3.onResize(function() {
 /// load all data
 var obsData = data[0],
     fiData = data[1], pdData = data[2],
-    adData = data[3], fdData = data[4];
+    adData = data[3], fdData = data[4],
+    tvData = data[5];
 
 /// load options
 var TIME = options.time,
@@ -73,7 +74,11 @@ var TIME = options.time,
     fdTitle = options.fd_title,
     fdSubtitle = options.fd_subtitle || subTitle,
     fdBarWidth = options.fd_bar_width || barWidth,
-    fdBarColor = options.fd_bar_color || barColor;
+    fdBarColor = options.fd_bar_color || barColor,
+    tvTitle = options.tv_title,
+    tvSubtitle = options.tv_subtitle || subTitle,
+    tvPointSize = options.tv_point_size || pointSize,
+    tvPointColor = options.tv_point_color || pointColor;
 
 /// for observation choice
 var observationIds = Object.keys(obsData);
@@ -110,6 +115,7 @@ if (!SHOW_SUBTITLE) {
   pdSubtitle = null;
   adSubtitle = null;
   fdSubtitle = null;
+  tvSubtitle = null;
 }
 
 /// for plot chosing
@@ -119,7 +125,8 @@ var notVisiblePlots = [{id:"BD", text:"Break Down [Local]"},
                        {id:"FI", text:"Feature Importance [Global]"},
                        {id:"PD", text:"Partial Dependency [Global]"},
                        {id:"AD", text:"Accumulated Dependency [Global]"},
-                       {id:"FD", text:"Feature Distribution [EDA]"}];
+                       {id:"FD", text:"Feature Distribution [EDA]"},
+                       {id:"TV", text:"Target vs Feature [EDA]"}];
 
 var visiblePlots = [];